rename ACAPS layout file
